Fix compilation warning by renaming forms.Mapping
